Refactor CTA components to dedicated directory structure
- Created cta directory with waitlist subcomponents:
  - waitlist-cta component
  - waitlist-creator-cta component
- Added comprehensive documentation with README files and usage examples
- Created consistent directory structure with index.ts exports
- Added backward compatibility re-exports with @deprecated annotations
- Updated COMPONENT-MOVEMENT-PLAN.md to track changes
