+++ src/components/ui/cta/index.ts
@@ -1,0 +1,1 @@
+export * from "./waitlist";
