+++ src/components/ui/cta/waitlist/index.ts
@@ -1,0 +1,2 @@
+export * from "./waitlist-cta";
+export * from "./waitlist-creator-cta";
